Split HslParkUpdater's code into different classes
